base de données supabase

+++ pages/welcome/onglets/factures.py
@@ -1174,7 +1174,7 @@ class Factures(ft.Container):
 
                         # si la remise est nulle
                         if int(self.edit_remise.value) == 0:
-                            mt_ir = int(mt_total*IR_VALUE[regime])
+                            mt_ir = mt_total*IR_VALUE[regime]
                             mt_nap = int(mt_total - mt_ir)
                             draw_simple_paragraph(
                                 f"IR:    {ajout_separateur(mt_ir)}",
@@ -1858,7 +1858,7 @@ class Factures(ft.Container):
 
                         # si la remise est nulle
                         if int(self.edit_remise.value) == 0:
-                            mt_ir = int(mt_total*IR_VALUE[regime])
+                            mt_ir = mt_total*IR_VALUE[regime]
                             mt_nap = int(mt_total - mt_ir)
                             draw_simple_paragraph(
                                 f"IR:    {ajout_separateur(mt_ir)}",
